requirements.txt
-i https://pypi.org/simple
agate==1.6.3
aiosqlite==0.19.0
alembic==1.10.3
anyio==3.6.2
apprise==1.3.0
asgi-lifespan==2.1.0
asyncpg==0.27.0
attrs==23.1.0
babel==2.12.1
betterproto==1.2.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.3.14
cryptography==40.0.2
dateparser==1.1.8
db-dtypes==1.1.1
dbt-bigquery==1.4.3
dbt-core==1.4.6
dbt-extractor==0.4.1
docker==6.0.1
fastapi==0.95.1
fsspec==2023.4.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-cloud-dataproc==5.4.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphql-core==3.2.3
greenlet==2.0.2
griffe==0.27.1
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hologram==0.0.16
hpack==4.0.0
httpcore==0.17.0
httplib2==0.22.0
httpx[http2]==0.24.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.5.0
isodate==0.6.1
jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kubernetes==26.1.0
leather==0.3.4
logbook==1.5.3
mako==1.2.4
markdown==3.4.3
markdown-it-py==2.2.0
markupsafe==2.1.2
mashumaro[msgpack]==3.3.1
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
numpy==1.24.2
oauthlib==3.2.2
orjson==3.8.10
packaging==23.1
pandas==1.5.2
pandas-gbq==0.18.1
parsedatetime==2.4
pathspec==0.10.3
pendulum==2.1.2
prefect==2.7.7
prefect-dbt[bigquery]==0.3.1
prefect-gcp[cloud_storage]==0.2.4
prefect-github==0.1.5
prefect-shell==0.1.5
proto-plus==1.22.2
protobuf==4.22.3
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydata-google-auth==1.7.0
pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pyyaml==6.0
readchar==4.0.5
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
setuptools==67.7.1
sgqlc==16.1
six==1.16.0
sniffio==1.3.0
sqlalchemy[asyncio]==1.4.47
sqlparse==0.4.3
starlette==0.26.1
stringcase==1.2.0
text-unidecode==1.3
toml==0.10.2
typer==0.7.0
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
websocket-client==1.5.1
werkzeug==2.2.3
zipp==3.15.0
